AI: let's start the project
Prompt:
We are going to be creating a Rust terminal based application. We are
going to use the ratatui library for rendering the TUI. The application
is going to be used a mathematical note taking application. I imagine
that the there will be a large text area where the user can type notes
(there should be line numbers in the UI) and those notes could include
equations with units, for example: (50 GiB / s) * 2 hours. The UI should
have a vertical divider around the 80% of the width where we will put
any results of equations that are found in that line. In the example
provided it would show 360,000 GiB. Let's focus on building just the UI
part now so that the user can type and see what they are typing. Then
once we are satisfied with that, we can move on to buildling the parsing
code that would detect any equations in the text.