AI: add unit tests
It really struggled with this one, i had to step in a few times and stop
it

prompt:

I think it would be useful to be able to unit test these capabilities.
Can you write some unit tests for the types of conversions that are
allowed. and while we are at, let's add unit tests for the expressions
too. Let's refactor so the unit tests operate just on the strings and
results

Ok great, now can you improve the code handling to make all the unit tests pass?

I think one of the problems we have is that the results are unit-less. I
think it would be better if all the results could support a unit too
(where applicable). So for example 1,000 GiB / 1 hour should just be
1,000 GiB/h. You'll probably want to change all the unit tests to expect
strings so you can include the unit that way