ai: add ability to evaluate expressions from command line without tty and more colors
prompt:
Now let's add the clap library so that we can provide some command line
features (such as the version), but in particular it would be great to
add support for something like "--" and after that it would treat that
line as if it were a line inside mathypad, so that it would parse it and
evaluate it. It should still print it using the same formatting (i.e.,
blue numbers) and then result. This would make it really useful for
one-shot equations

Now let's add more color (to both the interactive and non-interactive)
for any valid detected units