ai: review and cleanup
prompt:

Now I want you to review all the code we've written as if you were a senior engineer. I want you to find places that can
be optimized or improved. The goal is to add value, not make needless changes. In particular look out for parts of the
code that could be brittle or hard to understand.

I think there is quite a bit of dead code now (some marked with allow dead code). let's remove that. there
is no backwards compatibility to worry about since the application is not released, but let's make sure
nothing breaks