ai: add vim motion support
prompt:
Now let's add support for vim motion key emulation. Let's add support
for a normal and insert mode (insert should be the default), and after
pressing ESC it will switch to normal mode. In normal mode hjkl will be
able to move the cursor. Make sure to also make a small indicicator
(maybe in the bottom left, along the border) whenever the user is in
NORMAL mode