ai: restore the tests
This was quite bad, it conveniently removed all the unit tests during
the refactor.

I used the old git version of `src/main.rs` prior to the refactor and
gave it as context for how to restore the tests.

I honestly felt like i wanted a refund for that $13 refactor :P

prompt:
It is not acceptable that all the test were removed as part of the
refactoring. I have placed a filed called "old_main.rs" which contains
the version of the code and tests before the refactoring. Can you make
re-add the tests in their appropriate files to make sure we have test
coverage?