ai: write chumksy parser
This on was really painful. It spent so much time (and like $20 bucks)
trying to add support for variables with our existing parser
implementation but it was introducing regressions left and right.

So i told it to give up and write a chumksy parser version, and even
then i had to stop it a few times and give it some examples.

prompt:
My conclusion is that we are going to have to rewrite the parser to be much more powerful in
order to add support for variables. So I have reset all the changes you made. I think you
should start by re-implementing the parser using the chomsky rust library (chumsky = "0.10.1")
making sure all our existing tests and behavior still works.

I definitely want you to switch to using chumsky. Maybe you can try by
implementing it in another file? and starting from scratch on the parser
implementation?

I have reset the parser.rs file to how it was, that way the code will
compile again. Please continue implementing the parser using chumsky in
the separate file

Also don't forget you can get the chumsky docs here: https://
docs.rs/chumsky/0.10.1/chumsky/

I download an example chumsky implementation in the file
sample_chumsky.rs that you should use as reference