ai: fix finding mathematical expressions
this was running into an infinite loop, ended up going down a rabbit
hole where it was going to re-implement it with regex but luckily i
caught it and stopped, we eventually got to the where we wanted. glad i
had it write all those tests.

prompt:
There seems to be a bug with test_find_math_expression_complex and i think it's because
we are still using the old find_math_expression function which is not using the
chumksy parser. Can you fix that to use chumksy? Also start by considering whether we
need that function or not now that we have the chumsky parser

Why do we need regex at all? isn't the parser doing that already?

I actually think that "1 + 2)" should not evaluate to anything since it's an invalid expression