ai: change tokenizer to parse text
Better to have the tokenizer parse the text and then the evaluator
ignore it than the crazy hack that was find mathematical expression.
This one was tough again, had to do lots of stopping and re-prompting.
It was really getting all tripped up with this old implementation of
find mathematical expression.

prompt:
Would it make sense to have the parser also tokenize random text but
then the evaluator can ignore it? that way we can avoid needing to find
mathematical expressions?

Are you sure we even need the find_math_expression if we were to switch
to this method?

Yes let's break it down and implement this approach

Ok now I want you to fix the colon issue. I remember we had an issue
where the "," were somehow splitting, maybe there is something similar
with the colon?