ai: add github corner
i am amazed at how well this worked! i gave it an example from:
https://tholman.com/github-corners/

and it integrated it perfectly

prompt:
I have placed a file in web/example-corner.html with an example of a
cool way to show a link to the github repo. I want you to integrate that
into the website (with a link to the mathypad repo) on the top right
corner