ai: fix variables not re-evaluating dependent lines
prompt:
There seems to be a bug with variable, changing the value of one doesn't
seem to update expressions that use it. For example if I have
  "
  cars = 4
  cars * 2
  "
I would get 8 for the second line. But if i change the 4 to a 5, it
still stays at 8 for the second line.

I would like you to fix this