ai: add animations
prompt:
This is great. Now I think it would be delightful if there were some
animations. For example a cool fade animation on teh character whenever
we display the result of evaulating a line. can you add that?

I think the animation is too slow, can you make it faster? maybe 250 ms?