ai: write a release script
I tried giving it the bash script as prose and it worked!

prompt:
Can you write a small batch script that I can use for releases? I want
it to run "git cliff --bump" to generate the changelog, show me the
changes that are going to be in the latest version (The one i am
planning to release) and ask to confirm. if i confirm i want to run "git
cliff --bump -o" to write the changelog file, then i want it to edit the
Cargo.toml with the new version (the latest one in the change log). Then
i want it to run "cargo check" so it updates the cargo lock file. Then I
want to do a git commit with the message "no ai: v0.1.1" here the
"v0.1.1" is the correct version. finally i want it to create a git tag
with the version number as the tag and the contents of the changes as
the tag contents.