ai: confirmation dialog before quitting if not saved
prompt:
Now i think it would be useful to show a popup if the user is trying to
close and there are unsaved changes. it can show ctrl-s for save and
quit or ctrl-c to quit without saving.

There appears to be a bug whenever i start mathypad without a file, i
make some changes, try to quit, hit ctrl-s to save, type the filename,
hit ok. at that point my terminal ends up all messed up (this doesn't
happen if i first save, then exit; nor does it happen if you open with a
file). are we forgetting to reset the mode or something?