ai: add support for generating auto completion
prompt:
It would be great if on linux when you run mathypad and are typing the
next argument (for which file to open) that it would automatically
auto-complete (including tab support) for files that end in .pad (the
file extension of mathypad) in that directory. Can you add this?

How would this work for users that are not building mathypad directly?

Can you change the instructions to avoid having to write the completions
to a file? for example i see that starship suggests something like eval
$(starship init zsh) and i see atuin suggest: eval $(atuin init zsh) as
well

When i tested on zshrc, i get the following error "_mathypad:9: command
not found: _arguments"

That looks great. Although when i compile i get: warning:
mathypad@0.1.8: Copied completion: "mathypad.bash" warning:
mathypad@0.1.8: Copied completion: "mathypad.zsh" warning:
mathypad@0.1.8: Copied completion: "mathypad.fish" Can we remove that?
let's not support adding the completions when building, i think the eval
way is enough