removed non-generic rates
prompt:
I have a PR contribution from an outside author for mathypad. This is already checked out, and when I attempted to rebase it I hit merge conflicts. I want you to carefully spend time trying to understand the motivation behind the author's PR (by looking at the diff). And then carefully resolve the conflicts making sure to both preserve the intentions and behaviors, without sacrificing code quality.

Can you add some more unit tests that cover the new advanced rate expression support (make sure to include some non-simple path cases)

Ok so I think the reason we are having problems is that some of the tests still use the old (which we want to remove afterwards) way of specifying data rates. I think it may be worth it to update all the tests that deal with data rate units to use the new RateUnit rather than the old BytesPerSecond (and similar ones)

I think the display logic should be reworked to be able to function in the general case (rather than enumerating every possible permutation, which is bad). I want you to implement a working implementation, even if you have to allocate for now (we can always fix that later)

It is not great that all the previously static strings are now being called "to_string()" which needlessly allocates. Instead I want you to use a Cow type to avoid allocating in the cases where you don't have to

I think there is a bug, I tried computing "180 MiB/min in KiB/min" and i got the answer of "3 KiB/min" which is incorrect. It should be "184,320 KiB/min" since one 1 MiB = 1024 KiB. Can you add a test case for this exact case, and then fix the code.

Ok now that it is working with the generic RateUnit, i want you to remove all the old Unit types that can be represented with the new RateUnit. For example, we should have BytesPerSecond

That is great, but when I tried "180 MiB/min in KiB/hour" i got a
correct answer but in the wrong units. I just got "180 MiB/min" but i
wanted an answer in "KiB/hour" (hence the "in"). Add a new test for this
and fix it

Co-authored-by: Bernard Gallet <github@gallet.com>