ai: add support for copying expressions or results with double click
prompt:
Now I want you to make it so that when you double click on a result (or
a line on the pad) it will copy it to the user clipboard. This should
work on both MacOS and Linux. And it would be great to have a visual
indicator (maybe like a flash of the line containing the text that was
copied) to show the user that it was copied

That is almost right, i think there are some problems 1. the flashing
animation is too slow so it feels clunky 2. the flash shows for both the
expression line and the result line, and it should only show for the one
that i double clicked 3. i only seem to be able to copy the first line,
it didn't work for any other line