ai: improve the changelog generation
It really struggled with this one

prompt:
Ok now moving on to something else. We use "git cliff" to generate
change logs (for example "git cliff --bump" whenever we are going to
make a new release). However, I noticed that not all commits are being
included in the changelog. I want you to read about git cliff's
configurations: https://git-cliff.org/docs/configuration/ Then i want
you to update the current cliff.toml so that it will correctly include
all commits. And I want you to add support for commits that start with
"ai" to get a little robot emoji and be grouped together, whereas comits
that start with "no ai" to have a human emoji section. Feel free to use
"git cliff --bump" to see what the output of the changelog would be
until it is correct

Can you make it exclude the version bump commits (for example no need to
see "v0.1.9")?

I still see the version bumps (for example "- V0.1.9")

There are new lines missing in between the entries