ai: added ability to view changelog and save version
prompt:
Now I want you to add support for viewing the changelog from mathypad
with "--changelog" or "--whats-new". Mathypad should also write a small
file with the current version to "~/.mathypad". Not in this change, but
i will later want to build a welcome screen whenever you run a newer
version of mathypad with the latest changes

I ran it and saw the the version file got created

I think you should avoid all the build.rs code for generating the
changelog and instead just use the CHANGELOG.md that is already here in
the directory. Can you simplify all that?