ai: add a welcome message with the latest changes
so users of mathypad can know what's up!

prompt:
Ok now I want to add support for a welcome screen the first time you run
mathypad and it detects that the last time it was run it was a previous
version (by using the version file in ~/.mathypad). It should show a
popup screen with the changelog since the last version (it should be
scrollable if there is more than what can be displayed)

I manuall edited the version file and then opened mathypad but i didn't
see the welcome screen. are you sure that works? maybe we write the
version file before we check?

That worked! Can you add a scrollbar to the welcome screen so it's
obvious that there is more content below?

It appears that the last line of the changelog is cut off. Also the
bottom part of the welcome screen (with the text for scroll, enter or
esc to close) doesn't actually display at the bottom of the box. I think
it would look nicer if it did

Can you change the handling such that if there was no version, it will
also display the changes from the latest version?