ai: update website
prompt:
Ok now I want you to update the website given all the new functionality
that we've added. Make sure to not be too verbose or AI sounding, just
make sure the website is up to date with all the most useful
functionality

I think you can add some more examples of complex queries mathypad
supports (for example add one with line references). And maybe avoid the
ones with variable assignment?

That is getting better. Now i want you to take into account that the
audience for this is probably all software engineers, so make sure the
writing is adapted well for the audience.