ai: add support for vim command mode (e.g., :w, :wq, :cq)
prompt:
now i want you to extend vim support by adding a command mode support,
so that if in normal mode you type ':' then it should start typing below
in the command line. Users should be able to type commands such as ':q'
to quit or ':w' to save (it should trigger the save behavior as we
already have for saving). It should also support ':wq' to save and quit.
And saving should support writing to a file, for example ':w
some-file.pad'.

That is good! Except when I quit using :q it looks like it messed up my
terminal. Maybe we didn't handle raw mode properly when exiting?

Great, now add support for :q! and :cq to exit without saving and
without confirmation