ai: add support for currencies (e.g., $42 * 3)
prompt:
Now I want you to add support for currency as a unit. We want to make
sure that we preserve currency units (for example the pound symbol or
dollar symbol. Currenc units can be added with other of the same
currency, and can be multiplied or divided (for example $24 * 3 = $72)

I do not think you should allow division of different currencies.
Without the exchange rates we cannot give a correct answer, so make sure
divison is only allowed within the same currency