wip: ai: add editorialized release notes
prompt:
Ok now I want to add support for adding a free-form section to the
changelog whenever adding a new version. This should be in addition to
the full changelog that is currently generated by git cliff. But I would
like the ability to editorialize the changes. This will be particularly
useful for the welcome screen whenever there is an update, since the git
commit messages might not be descriptive enough for all users. Can you
add a step in the release script for this, and then make sure it gets
added to the changelog (and previous ones should be preserved whenever
we-regenerate the change log!), and finally make sure that the welcome
screen would also include the editorialized changelog as part of the
popup

You should also make sure that the editorialized part ends up on the
github release too. Continue