ai: fix currrency rate conversions (e.g., $500 / year to $/month)
prompt:
Great, thanks. I want you to add support for expressions like "$5/month
to $/year" which should evaluate to "$60/year"

I dont think that failed currency conversion tests should fall back to
their original value, that could be very confusing to the user. If you
try to do "$5/hr to €/hr" it should fail and not evaulate to anything.
That should also be true for "5 MB to QPS" that shouldn't evaluate to
anything. But this should only be the case if there was an
operator/operation. If there is just text afterards it should continue
to evaluate, for example "$5 is how much the ice cream cost" should
evaluate to "$5"