ai: add panic handler in case mathypad crashes (will save error logs and save current file)
WIP because i still need to test this. i should be able to make a
branch, remove the commit that fixed utf8 handling, and paste euros
again to repro and make sure this fixed worked

prompt:
One thing that was painful is that when it panics it was just rendering
the stack trace in the mathypad app which is hard to read, and it seems
like it kept my terminal in raw mode which made it not possible to use.
Can you fix both of those things? Maybe make sure that any errors get
written to an error file instead of draw on the mathypad TUI? and make
sure that even if we panic we still properly clean up the app and leave
the terminal in a good mode

Can you do it but without using chrono? that is a big library with lots
of dependencies that doesn't seem worth it for this. a simple unix
timestamp should suffice instead of chrono