ai: fix line references
prompt:
Ok that is good. It appears that in the egui version the line references
don't update in the text. For example if I had line3 and hit enter in
line1, i expected it to become line4 in the text. We had already solved
this for the tui version, but it seems to have been broken. Make sure to
fix it in a way that will work for both the TUI and the egui versions