ai: fix color scheme
prompt:
That is great it seems to have fixed the flickering, but the colors for
the native and web egui versions don't look great against the white
background. Maybe you can have separate colors for the egui version
which has a light background? and leave the colors for the TUI the same

i still don't like the light-mode colors, for example line references
are too light and numbers are too light too.

[Image #1] they are still hard to read. For exmaple operators are almost
impossible to read. Try to come up with an elegant light color scheme
please