ai: Fix safe area in PWA mode
prompt:
Now there is another problem on the mobile version of the web egui. When
i added it to my home screen and opened it as an app, the top of the
application renders conflictingly with the dynamic island of my iphone.
Is there a way to have a safe area where it won't render to make sure I
can see everything?