ai: add support for sum_above()
prompt:
Can you now add support for a function called sum_above() which takes no
arguments and will sum the results of the lines above (assuming they can
be summed following all the existing rules about what can be summed). We
need to make sure that it works even if there are multiple lines with
sum_above() in an efficient way. Make sure that it is highlighted the
same way our other functions are (for example sqrt()). And write tests
for it ensuring the correctness without introducing any regressions

There is no need for sum_above() in one shot mode, since there can not
be any results above. So make sure we aren't trying to implement it for
that

There to seems to be a slight bug with it. If i have a mathypad with
sum_above and i ented a line above and type an expression, it doesn't
immediately update the line that has the sum_above() until i hit enter.
it would be better that it would dynamically update as soon as I make
changes to any of the lines above

That seems to be working correctly, but I noticed that sum_above()
doesnt have the same syntax highlighting as sqrt(). Can you fix that?